// src/app/Core/models/institucion.ts
export class Institucion {
    insId: number = 0;
    insNombre: string = '';
    insDireccion: string = '';
    insLogo: string=''; 
    insLogoGrande: string='';
    insLogosec: string='';
    insLogogob: string='';
    insLogosecretariaedu: string='';
    insLogocert: string='';
    insEstado: boolean = true;
}